requirements.txt
biopython
pandas
streamlit